Remove jq dependency, add threatstack and salt
For text-only output, this script now runs entirely in bash.

Also added the current services and checks if threatstack and
salt-minion are running for systemd systems.

Also updates the "admin users" report to include the sudo group, not
just the wheel group.

For json output, this script now depends on Python and bash >= 4.1
rather than jq. It also does not provide a method of requesting json
rather than text from the command line; editing the main() function is
necessary at this time.

Also note the "os_distro" key in the json output has been renamed to
"distro_full" for reasons of internal consistency. Since this json
output was historically parsed by human eyes, this is not expected to be
a breaking change.

Technically bash 4.0 is the minimum requirement for text-only output,
but the increase in logical complexity to both determine and communicate
that seemed less helpful.